Extract schema validation tests into common tests
